feat: [정다운] 상품 목록 페이지에서 리뷰 평균 평점을 소수점 단위로 확인할 수 있도록 변경 [SDY-124]

+++ src/main/java/com/dyes/backend/domain/product/service/user/UserProductServiceImpl.java
@@ -349,7 +349,7 @@ public class UserProductServiceImpl implements UserProductService {
 
         List<Review> reviewList = reviewRepository.findAllByProduct(product);
         int totalReviewCount = 0;
-        int averageRating = 0;
+        double averageRating = 0;
         int totalRating = 0;
         for(Review review : reviewList) {
             totalReviewCount = totalReviewCount + 1;
